+++ dropwizard-core/src/main/java/com/codahale/dropwizard/server/AbstractServerFactory.java
@@ -2,7 +2,7 @@ package com.codahale.dropwizard.server;
 
 import com.codahale.dropwizard.jersey.jackson.JacksonMessageBodyProvider;
 import com.codahale.dropwizard.jersey.setup.JerseyEnvironment;
-import com.codahale.dropwizard.jetty.GzipHandlerFactory;
+import com.codahale.dropwizard.jetty.GzipFilterFactory;
 import com.codahale.dropwizard.jetty.NonblockingServletHolder;
 import com.codahale.dropwizard.jetty.RequestLogFactory;
 import com.codahale.dropwizard.lifecycle.setup.LifecycleEnvironment;
@@ -27,6 +27,7 @@ import org.eclipse.jetty.server.Handler;
 import org.eclipse.jetty.server.Server;
 import org.eclipse.jetty.server.handler.ErrorHandler;
 import org.eclipse.jetty.server.handler.RequestLogHandler;
+import org.eclipse.jetty.servlet.FilterHolder;
 import org.eclipse.jetty.servlet.ServletContextHandler;
 import org.eclipse.jetty.util.BlockingArrayQueue;
 import org.eclipse.jetty.util.thread.ThreadPool;
@@ -62,7 +63,7 @@ import java.util.regex.Pattern;
  *     <tr>
  *         <td>{@code gzip}</td>
  *         <td></td>
- *         <td>The {@link GzipHandlerFactory GZIP} configuration.</td>
+ *         <td>The {@link GzipFilterFactory GZIP} configuration.</td>
  *     </tr>
  *     <tr>
  *         <td>{@code maxThreads}</td>
@@ -99,7 +100,7 @@ public abstract class AbstractServerFactory implements ServerFactory {
 
     @Valid
     @NotNull
-    private GzipHandlerFactory gzip = new GzipHandlerFactory();
+    private GzipFilterFactory gzip = new GzipFilterFactory();
 
     @Min(2)
     private int maxThreads = 1024;
@@ -129,12 +130,12 @@ public abstract class AbstractServerFactory implements ServerFactory {
     }
 
     @JsonProperty("gzip")
-    public GzipHandlerFactory getGzipHandlerFactory() {
+    public GzipFilterFactory getGzipFilterFactory() {
         return gzip;
     }
 
     @JsonProperty("gzip")
-    public void setGzipHandlerFactory(GzipHandlerFactory gzip) {
+    public void setGzipFilterFactory(GzipFilterFactory gzip) {
         this.gzip = gzip;
     }
 
@@ -194,6 +195,10 @@ public abstract class AbstractServerFactory implements ServerFactory {
                                        @Nullable ServletContainer jerseyContainer,
                                        MetricRegistry metricRegistry) {
         handler.addFilter(ThreadNameFilter.class, "/*", EnumSet.of(DispatcherType.REQUEST));
+        if (gzip.isEnabled()) {
+            final FilterHolder holder = new FilterHolder(gzip.build());
+            handler.addFilter(holder, "/*", EnumSet.allOf(DispatcherType.class));
+        }
         if (jerseyContainer != null) {
             jersey.addProvider(new JacksonMessageBodyProvider(objectMapper, validator));
             handler.addServlet(new NonblockingServletHolder(jerseyContainer), jersey.getUrlPattern());
@@ -222,15 +227,14 @@ public abstract class AbstractServerFactory implements ServerFactory {
         return server;
     }
 
-    protected Handler addGzipAndRequestLog(Handler handler, String name) {
-        final Handler gzipHandler = gzip.wrapHandler(handler);
+    protected Handler addRequestLog(Handler handler, String name) {
         if (requestLog.isEnabled()) {
             final RequestLogHandler requestLogHandler = new RequestLogHandler();
             requestLogHandler.setRequestLog(requestLog.build(name));
-            requestLogHandler.setHandler(gzipHandler);
+            requestLogHandler.setHandler(handler);
             return requestLogHandler;
         }
-        return gzipHandler;
+        return handler;
     }
 
     protected void printBanner(String name) {